pyproject.toml
[project]
name = "table-migrator"
version = "0.1.0"
description = "Reads data from a source database DSN and writes to a destination database DSN"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyhive[hive-pure-sasl]>=0.7.0",
    "sqlalchemy-hana>=2.8.0",
    "pandas>=2.2.3",
    "pydantic-settings>=2.6.1",
    "loguru>=0.7.2",
    "rich>=13.9.4",
    "rich-argparse>=1.6.0",
]
authors = [{ name = "Baeg-il Kim", email = "dt101665@mobis-partners.com" }]
license = { text = "MIT" }

[project.scripts]
table_migrate = "table_migrate:main"

[tool.uv]
dev-dependencies = ["ptpython>=3.0.29"]

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ['table_migrate.py']
source-includes = ["scripts/", "tests/"]

[dependency-groups]
dev = [
    "ptpython>=3.0.29",
    "pudb>=2024.1.3",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
